Identify posterior mean differences

Compute a criterion based on Credible Intervals (CI) to determine whether the posterior t-distributions of groups should be considered different enough to deserve further examination. Two groups are considered probably 'distinct' if the Credible Interval of level CI_level of their respective posterior t-distributions do not overlap.

Usage

identify_diff(posterior, CI_level = 0.05, nb_sample = 1000)

Arguments

posterior

A tibble, typically coming from a posterior_mean() function, containing the parameters of the multivariate posterior t-distributions for the mean of the considered groups and draws for each peptide.

CI_level

A number, defining the order of quantile chosen to assess differences between groups.

nb_sample

A number (optional), indicating the number of samples to draw from the posteriors for computing mean and credible intervals . Only used if posterior is multivariate, typically coming from a multi_posterior_mean() function.

Value

A tibble, indicating which peptides and groups seem to be different

Examples

TRUE
#> [1] TRUE